Add fetch wrapper that creates network logs
